@@ -213,9 +213,10 @@ public class BlockAzalea extends BlockFlowable {
         if (ev.isCancelled() || !success) {
             return;
         }
-        this.level.setBlock(this.getLocation(), Block.get(BlockID.AIR));
+
         for(Block block : ev.getBlockList()) {
             this.level.setBlock(block, block);
         }
+        this.level.setBlock(this, Block.get(LOG));
     }
 }
