@@ -646,153 +646,153 @@ public interface BlockID {
     @Since("1.4.0.0-PN") @PowerNukkitOnly int CRACKED_NETHER_BRICKS = 558;
     @Since("1.4.0.0-PN") @PowerNukkitOnly int QUARTZ_BRICKS = 559;
     // 560 Special block: minecraft:unknown
-    @Since("FUTURE") @PowerNukkitOnly int POWDER_SNOW = 561;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POWDER_SNOW = 561;
     @Since("FUTURE") @PowerNukkitOnly int SCULK_SENSOR = 562;
-    @Since("FUTURE") @PowerNukkitOnly int POINTED_DRIPSTONE = 563;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POINTED_DRIPSTONE = 563;
     // 564 (unused)
     // 565 (unused)
-    @Since("FUTURE") @PowerNukkitOnly int COPPER_ORE = 566;
-    @Since("FUTURE") @PowerNukkitOnly int LIGHTNING_ROD = 567;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COPPER_ORE = 566;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIGHTNING_ROD = 567;
     // 568 (unused)
     // 569 (unused)
     // 570 (unused)
     // 571 (unused)
-    @Since("FUTURE") @PowerNukkitOnly int DRIPSTONE_BLOCK = 572;
-    @Since("FUTURE") @PowerNukkitOnly int DIRT_WITH_ROOTS = 573;
-    @Since("FUTURE") @PowerNukkitOnly int HANGING_ROOTS = 574;
-    @Since("FUTURE") @PowerNukkitOnly int MOSS_BLOCK = 575;
-    @Since("FUTURE") @PowerNukkitOnly int SPORE_BLOSSOM = 576;
-    @Since("FUTURE") @PowerNukkitOnly int CAVE_VINES = 577;
-    @Since("FUTURE") @PowerNukkitOnly int BIG_DRIPLEAF = 578;
-    @Since("FUTURE") @PowerNukkitOnly int AZALEA_LEAVES = 579;
-    @Since("FUTURE") @PowerNukkitOnly int AZALEA_LEAVES_FLOWERED = 580;
-    @Since("FUTURE") @PowerNukkitOnly int CALCITE = 581;
-    @Since("FUTURE") @PowerNukkitOnly int AMETHYST_BLOCK = 582;
-    @Since("FUTURE") @PowerNukkitOnly int BUDDING_AMETHYST = 583;
-    @Since("FUTURE") @PowerNukkitOnly int AMETHYST_CLUSTER = 584;
-    @Since("FUTURE") @PowerNukkitOnly int LARGE_AMETHYST_BUD = 585;
-    @Since("FUTURE") @PowerNukkitOnly int MEDIUM_AMETHYST_BUD = 586;
-    @Since("FUTURE") @PowerNukkitOnly int SMALL_AMETHYST_BUD = 587;
-    @Since("FUTURE") @PowerNukkitOnly int TUFF = 588;
-    @Since("FUTURE") @PowerNukkitOnly int TINTED_GLASS = 589;
-    @Since("FUTURE") @PowerNukkitOnly int MOSS_CARPET = 590;
-    @Since("FUTURE") @PowerNukkitOnly int SMALL_DRIPLEAF_BLOCK = 591;
-    @Since("FUTURE") @PowerNukkitOnly int AZALEA = 592;
-    @Since("FUTURE") @PowerNukkitOnly int FLOWERING_AZALEA = 593;
-    @Since("FUTURE") @PowerNukkitOnly int GLOW_FRAME = 594;
-    @Since("FUTURE") @PowerNukkitOnly int COPPER_BLOCK = 595;
-    @Since("FUTURE") @PowerNukkitOnly int EXPOSED_COPPER = 596;
-    @Since("FUTURE") @PowerNukkitOnly int WEATHERED_COPPER = 597;
-    @Since("FUTURE") @PowerNukkitOnly int OXIDIZED_COPPER = 598;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_COPPER = 599;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_EXPOSED_COPPER = 600;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_WEATHERED_COPPER = 601;
-    @Since("FUTURE") @PowerNukkitOnly int CUT_COPPER = 602;
-    @Since("FUTURE") @PowerNukkitOnly int EXPOSED_CUT_COPPER = 603;
-    @Since("FUTURE") @PowerNukkitOnly int WEATHERED_CUT_COPPER = 604;
-    @Since("FUTURE") @PowerNukkitOnly int OXIDIZED_CUT_COPPER = 605;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_CUT_COPPER = 606;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_EXPOSED_CUT_COPPER = 607;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_WEATHERED_CUT_COPPER = 608;
-    @Since("FUTURE") @PowerNukkitOnly int CUT_COPPER_STAIRS = 609;
-    @Since("FUTURE") @PowerNukkitOnly int EXPOSED_CUT_COPPER_STAIRS = 610;
-    @Since("FUTURE") @PowerNukkitOnly int WEATHERED_CUT_COPPER_STAIRS = 611;
-    @Since("FUTURE") @PowerNukkitOnly int OXIDIZED_CUT_COPPER_STAIRS = 612;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_CUT_COPPER_STAIRS = 613;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_EXPOSED_CUT_COPPER_STAIRS = 614;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_WEATHERED_CUT_COPPER_STAIRS = 615;
-    @Since("FUTURE") @PowerNukkitOnly int CUT_COPPER_SLAB = 616;
-    @Since("FUTURE") @PowerNukkitOnly int EXPOSED_CUT_COPPER_SLAB = 617;
-    @Since("FUTURE") @PowerNukkitOnly int WEATHERED_CUT_COPPER_SLAB = 618;
-    @Since("FUTURE") @PowerNukkitOnly int OXIDIZED_CUT_COPPER_SLAB = 619;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_CUT_COPPER_SLAB = 620;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_EXPOSED_CUT_COPPER_SLAB = 621;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_WEATHERED_CUT_COPPER_SLAB = 622;
-    @Since("FUTURE") @PowerNukkitOnly int DOUBLE_CUT_COPPER_SLAB = 623;
-    @Since("FUTURE") @PowerNukkitOnly int EXPOSED_DOUBLE_CUT_COPPER_SLAB = 624;
-    @Since("FUTURE") @PowerNukkitOnly int WEATHERED_DOUBLE_CUT_COPPER_SLAB = 625;
-    @Since("FUTURE") @PowerNukkitOnly int OXIDIZED_DOUBLE_CUT_COPPER_SLAB = 626;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_DOUBLE_CUT_COPPER_SLAB = 627;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_EXPOSED_DOUBLE_CUT_COPPER_SLAB = 628;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_WEATHERED_DOUBLE_CUT_COPPER_SLAB = 629;
-    @Since("FUTURE") @PowerNukkitOnly int CAVE_VINES_BODY_WITH_BERRIES = 630;
-    @Since("FUTURE") @PowerNukkitOnly int CAVE_VINES_HEAD_WITH_BERRIES = 631;
-    @Since("FUTURE") @PowerNukkitOnly int SMOOTH_BASALT = 632;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE = 633;
-    @Since("FUTURE") @PowerNukkitOnly int COBBLED_DEEPSLATE = 634;
-    @Since("FUTURE") @PowerNukkitOnly int COBBLED_DEEPSLATE_SLAB = 635;
-    @Since("FUTURE") @PowerNukkitOnly int COBBLED_DEEPSLATE_STAIRS = 636;
-    @Since("FUTURE") @PowerNukkitOnly int COBBLED_DEEPSLATE_WALL = 637;
-    @Since("FUTURE") @PowerNukkitOnly int POLISHED_DEEPSLATE = 638;
-    @Since("FUTURE") @PowerNukkitOnly int POLISHED_DEEPSLATE_SLAB = 639;
-    @Since("FUTURE") @PowerNukkitOnly int POLISHED_DEEPSLATE_STAIRS = 640;
-    @Since("FUTURE") @PowerNukkitOnly int POLISHED_DEEPSLATE_WALL = 641;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_TILES = 642;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_TILE_SLAB = 643;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_TILE_STAIRS = 644;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_TILE_WALL = 645;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_BRICKS = 646;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_BRICK_SLAB = 647;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_BRICK_STAIRS = 648;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_BRICK_WALL = 649;
-    @Since("FUTURE") @PowerNukkitOnly int CHISELED_DEEPSLATE = 650;
-    @Since("FUTURE") @PowerNukkitOnly int COBBLED_DEEPSLATE_DOUBLE_SLAB = 651;
-    @Since("FUTURE") @PowerNukkitOnly int POLISHED_DEEPSLATE_DOUBLE_SLAB = 652;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_TILE_DOUBLE_SLAB = 653;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_BRICK_DOUBLE_SLAB = 654;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_LAPIS_ORE = 655;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_IRON_ORE = 656;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_GOLD_ORE = 657;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_REDSTONE_ORE = 658;
-    @Since("FUTURE") @PowerNukkitOnly int LIT_DEEPSLATE_REDSTONE_ORE = 659;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_DIAMOND_ORE = 660;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_COAL_ORE = 661;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_EMERALD_ORE = 662;
-    @Since("FUTURE") @PowerNukkitOnly int DEEPSLATE_COPPER_ORE = 663;
-    @Since("FUTURE") @PowerNukkitOnly int CRACKED_DEEPSLATE_TILES = 664;
-    @Since("FUTURE") @PowerNukkitOnly int CRACKED_DEEPSLATE_BRICKS = 665;
-    @Since("FUTURE") @PowerNukkitOnly int GLOW_LICHEN = 666;
-    @Since("FUTURE") @PowerNukkitOnly int CANDLE = 667;
-    @Since("FUTURE") @PowerNukkitOnly int WHITE_CANDLE = 668;
-    @Since("FUTURE") @PowerNukkitOnly int ORANGE_CANDLE = 669;
-    @Since("FUTURE") @PowerNukkitOnly int MAGENTA_CANDLE = 670;
-    @Since("FUTURE") @PowerNukkitOnly int LIGHT_BLUE_CANDLE = 671;
-    @Since("FUTURE") @PowerNukkitOnly int YELLOW_CANDLE = 672;
-    @Since("FUTURE") @PowerNukkitOnly int LIME_CANDLE = 673;
-    @Since("FUTURE") @PowerNukkitOnly int PINK_CANDLE = 674;
-    @Since("FUTURE") @PowerNukkitOnly int GRAY_CANDLE = 675;
-    @Since("FUTURE") @PowerNukkitOnly int LIGHT_GRAY_CANDLE = 676;
-    @Since("FUTURE") @PowerNukkitOnly int CYAN_CANDLE = 677;
-    @Since("FUTURE") @PowerNukkitOnly int PURPLE_CANDLE = 678;
-    @Since("FUTURE") @PowerNukkitOnly int BLUE_CANDLE = 679;
-    @Since("FUTURE") @PowerNukkitOnly int BROWN_CANDLE = 680;
-    @Since("FUTURE") @PowerNukkitOnly int GREEN_CANDLE = 681;
-    @Since("FUTURE") @PowerNukkitOnly int RED_CANDLE = 682;
-    @Since("FUTURE") @PowerNukkitOnly int BLACK_CANDLE = 683;
-    @Since("FUTURE") @PowerNukkitOnly int CANDLE_CAKE = 684;
-    @Since("FUTURE") @PowerNukkitOnly int WHITE_CANDLE_CAKE = 685;
-    @Since("FUTURE") @PowerNukkitOnly int ORANGE_CANDLE_CAKE = 686;
-    @Since("FUTURE") @PowerNukkitOnly int MAGENTA_CANDLE_CAKE = 687;
-    @Since("FUTURE") @PowerNukkitOnly int LIGHT_BLUE_CANDLE_CAKE = 688;
-    @Since("FUTURE") @PowerNukkitOnly int YELLOW_CANDLE_CAKE = 689;
-    @Since("FUTURE") @PowerNukkitOnly int LIME_CANDLE_CAKE = 690;
-    @Since("FUTURE") @PowerNukkitOnly int PINK_CANDLE_CAKE = 691;
-    @Since("FUTURE") @PowerNukkitOnly int GRAY_CANDLE_CAKE = 692;
-    @Since("FUTURE") @PowerNukkitOnly int LIGHT_GRAY_CANDLE_CAKE = 693;
-    @Since("FUTURE") @PowerNukkitOnly int CYAN_CANDLE_CAKE = 694;
-    @Since("FUTURE") @PowerNukkitOnly int PURPLE_CANDLE_CAKE = 695;
-    @Since("FUTURE") @PowerNukkitOnly int BLUE_CANDLE_CAKE = 696;
-    @Since("FUTURE") @PowerNukkitOnly int BROWN_CANDLE_CAKE = 697;
-    @Since("FUTURE") @PowerNukkitOnly int GREEN_CANDLE_CAKE = 698;
-    @Since("FUTURE") @PowerNukkitOnly int RED_CANDLE_CAKE = 699;
-    @Since("FUTURE") @PowerNukkitOnly int BLACK_CANDLE_CAKE = 700;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_OXIDIZED_COPPER = 701;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_OXIDIZED_CUT_COPPER = 702;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_OXIDIZED_CUT_COPPER_STAIRS = 703;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_OXIDIZED_CUT_COPPER_SLAB = 704;
-    @Since("FUTURE") @PowerNukkitOnly int WAXED_OXIDIZED_DOUBLE_CUT_COPPER_SLAB = 705;
-    @Since("FUTURE") @PowerNukkitOnly int RAW_IRON_BLOCK = 706;
-    @Since("FUTURE") @PowerNukkitOnly int RAW_COPPER_BLOCK = 707;
-    @Since("FUTURE") @PowerNukkitOnly int RAW_GOLD_BLOCK = 708;
-    @Since("FUTURE") @PowerNukkitOnly int INFESTED_DEEPSLATE = 709;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DRIPSTONE_BLOCK = 572;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DIRT_WITH_ROOTS = 573;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int HANGING_ROOTS = 574;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int MOSS_BLOCK = 575;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int SPORE_BLOSSOM = 576;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CAVE_VINES = 577;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BIG_DRIPLEAF = 578;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int AZALEA_LEAVES = 579;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int AZALEA_LEAVES_FLOWERED = 580;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CALCITE = 581;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int AMETHYST_BLOCK = 582;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BUDDING_AMETHYST = 583;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int AMETHYST_CLUSTER = 584;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LARGE_AMETHYST_BUD = 585;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int MEDIUM_AMETHYST_BUD = 586;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int SMALL_AMETHYST_BUD = 587;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int TUFF = 588;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int TINTED_GLASS = 589;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int MOSS_CARPET = 590;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int SMALL_DRIPLEAF_BLOCK = 591;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int AZALEA = 592;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int FLOWERING_AZALEA = 593;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int GLOW_FRAME = 594;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COPPER_BLOCK = 595;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int EXPOSED_COPPER = 596;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WEATHERED_COPPER = 597;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int OXIDIZED_COPPER = 598;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_COPPER = 599;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_EXPOSED_COPPER = 600;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_WEATHERED_COPPER = 601;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CUT_COPPER = 602;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int EXPOSED_CUT_COPPER = 603;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WEATHERED_CUT_COPPER = 604;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int OXIDIZED_CUT_COPPER = 605;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_CUT_COPPER = 606;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_EXPOSED_CUT_COPPER = 607;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_WEATHERED_CUT_COPPER = 608;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CUT_COPPER_STAIRS = 609;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int EXPOSED_CUT_COPPER_STAIRS = 610;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WEATHERED_CUT_COPPER_STAIRS = 611;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int OXIDIZED_CUT_COPPER_STAIRS = 612;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_CUT_COPPER_STAIRS = 613;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_EXPOSED_CUT_COPPER_STAIRS = 614;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_WEATHERED_CUT_COPPER_STAIRS = 615;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CUT_COPPER_SLAB = 616;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int EXPOSED_CUT_COPPER_SLAB = 617;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WEATHERED_CUT_COPPER_SLAB = 618;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int OXIDIZED_CUT_COPPER_SLAB = 619;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_CUT_COPPER_SLAB = 620;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_EXPOSED_CUT_COPPER_SLAB = 621;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_WEATHERED_CUT_COPPER_SLAB = 622;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DOUBLE_CUT_COPPER_SLAB = 623;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int EXPOSED_DOUBLE_CUT_COPPER_SLAB = 624;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WEATHERED_DOUBLE_CUT_COPPER_SLAB = 625;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int OXIDIZED_DOUBLE_CUT_COPPER_SLAB = 626;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_DOUBLE_CUT_COPPER_SLAB = 627;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_EXPOSED_DOUBLE_CUT_COPPER_SLAB = 628;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_WEATHERED_DOUBLE_CUT_COPPER_SLAB = 629;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CAVE_VINES_BODY_WITH_BERRIES = 630;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CAVE_VINES_HEAD_WITH_BERRIES = 631;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int SMOOTH_BASALT = 632;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE = 633;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COBBLED_DEEPSLATE = 634;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COBBLED_DEEPSLATE_SLAB = 635;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COBBLED_DEEPSLATE_STAIRS = 636;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COBBLED_DEEPSLATE_WALL = 637;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POLISHED_DEEPSLATE = 638;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POLISHED_DEEPSLATE_SLAB = 639;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POLISHED_DEEPSLATE_STAIRS = 640;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POLISHED_DEEPSLATE_WALL = 641;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_TILES = 642;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_TILE_SLAB = 643;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_TILE_STAIRS = 644;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_TILE_WALL = 645;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_BRICKS = 646;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_BRICK_SLAB = 647;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_BRICK_STAIRS = 648;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_BRICK_WALL = 649;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CHISELED_DEEPSLATE = 650;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int COBBLED_DEEPSLATE_DOUBLE_SLAB = 651;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int POLISHED_DEEPSLATE_DOUBLE_SLAB = 652;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_TILE_DOUBLE_SLAB = 653;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_BRICK_DOUBLE_SLAB = 654;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_LAPIS_ORE = 655;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_IRON_ORE = 656;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_GOLD_ORE = 657;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_REDSTONE_ORE = 658;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIT_DEEPSLATE_REDSTONE_ORE = 659;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_DIAMOND_ORE = 660;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_COAL_ORE = 661;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_EMERALD_ORE = 662;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int DEEPSLATE_COPPER_ORE = 663;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CRACKED_DEEPSLATE_TILES = 664;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CRACKED_DEEPSLATE_BRICKS = 665;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int GLOW_LICHEN = 666;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CANDLE = 667;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WHITE_CANDLE = 668;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int ORANGE_CANDLE = 669;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int MAGENTA_CANDLE = 670;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIGHT_BLUE_CANDLE = 671;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int YELLOW_CANDLE = 672;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIME_CANDLE = 673;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int PINK_CANDLE = 674;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int GRAY_CANDLE = 675;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIGHT_GRAY_CANDLE = 676;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CYAN_CANDLE = 677;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int PURPLE_CANDLE = 678;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BLUE_CANDLE = 679;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BROWN_CANDLE = 680;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int GREEN_CANDLE = 681;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int RED_CANDLE = 682;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BLACK_CANDLE = 683;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CANDLE_CAKE = 684;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WHITE_CANDLE_CAKE = 685;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int ORANGE_CANDLE_CAKE = 686;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int MAGENTA_CANDLE_CAKE = 687;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIGHT_BLUE_CANDLE_CAKE = 688;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int YELLOW_CANDLE_CAKE = 689;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIME_CANDLE_CAKE = 690;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int PINK_CANDLE_CAKE = 691;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int GRAY_CANDLE_CAKE = 692;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int LIGHT_GRAY_CANDLE_CAKE = 693;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int CYAN_CANDLE_CAKE = 694;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int PURPLE_CANDLE_CAKE = 695;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BLUE_CANDLE_CAKE = 696;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BROWN_CANDLE_CAKE = 697;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int GREEN_CANDLE_CAKE = 698;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int RED_CANDLE_CAKE = 699;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int BLACK_CANDLE_CAKE = 700;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_OXIDIZED_COPPER = 701;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_OXIDIZED_CUT_COPPER = 702;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_OXIDIZED_CUT_COPPER_STAIRS = 703;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_OXIDIZED_CUT_COPPER_SLAB = 704;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int WAXED_OXIDIZED_DOUBLE_CUT_COPPER_SLAB = 705;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int RAW_IRON_BLOCK = 706;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int RAW_COPPER_BLOCK = 707;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int RAW_GOLD_BLOCK = 708;
+    @Since("1.6.0.0-PNX") @PowerNukkitOnly int INFESTED_DEEPSLATE = 709;
 }
