@@ -516,7 +516,7 @@ public class Item implements Cloneable, BlockID, ItemID {
                 // TODO Remove this when the support is added to these blocks
                 String[] stateParts = blockStateId.split(";", 2);
                 Integer blockId = BlockStateRegistry.getBlockId(stateParts[0]);
-                if (blockId != null && blockId > BlockID.QUARTZ_BRICKS) {
+                if (blockId != null && blockId > BlockID.INFESTED_DEEPSLATE) {
                     return Item.getBlock(BlockID.AIR);
                 }
 
