@@ -369,7 +369,7 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
                 return h;
             }
         }
-        for (int y = isOverWorld() ? 319 : 255; y >= 0; --y) {
+        for (int y = isOverWorld() ? 319 : 255; y >= (isOverWorld() ? -64 : 0); --y) {
             if (getBlockId(x, y, z) != 0x00) {
                 this.setHeightMap(x, z, y);
                 return y;
@@ -406,7 +406,7 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
             this.tileList = new Int2ObjectOpenHashMap<>();
         }
         this.tiles.put(blockEntity.getId(), blockEntity);
-        int index = ((blockEntity.getFloorZ() & 0x0f) << 12) | ((blockEntity.getFloorX() & 0x0f) << 8) | (blockEntity.getFloorY() & 0xff);
+        int index = ((blockEntity.getFloorZ() & 0x0f) << 16) | ((blockEntity.getFloorX() & 0x0f) << 12) | (ensureY(blockEntity.getFloorY()) + 64);
         if (this.tileList.containsKey(index) && !this.tileList.get(index).equals(blockEntity)) {
             BlockEntity entity = this.tileList.get(index);
             this.tiles.remove(entity.getId());
@@ -422,7 +422,7 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
     public void removeBlockEntity(BlockEntity blockEntity) {
         if (this.tiles != null) {
             this.tiles.remove(blockEntity.getId());
-            int index = ((blockEntity.getFloorZ() & 0x0f) << 12) | ((blockEntity.getFloorX() & 0x0f) << 8) | (blockEntity.getFloorY() & 0xff);
+            int index = ((blockEntity.getFloorZ() & 0x0f) << 16) | ((blockEntity.getFloorX() & 0x0f) << 12) | (ensureY(blockEntity.getFloorY()) + 64);
             this.tileList.remove(index);
             if (this.isInit) {
                 this.setChanged();
@@ -447,7 +447,7 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
 
     @Override
     public BlockEntity getTile(int x, int y, int z) {
-        return this.tileList != null ? this.tileList.get((z << 12) | (x << 8) | y) : null;
+        return this.tileList != null ? this.tileList.get((z << 16) | (x << 12) | (y + 64)) : null;
     }
 
     @Override
@@ -730,5 +730,15 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
         }
         
         return results.stream();
+    }
+
+    @PowerNukkitOnly
+    @Since("1.6.0.0-PNX")
+    private int ensureY(final int y) {
+        if (isOverWorld()) {
+            return Math.max(Math.min(y, 319), -64);
+        } else {
+            return y & 0xff;
+        }
     }
 }
